+++ lib/pry-nav.rb
@@ -5,3 +5,7 @@ require 'pry-nav/tracer'
 
 # Optionally load pry-remote monkey patches
 require 'pry-nav/pry_remote_ext' if defined? PryRemote
+
+module PryNav
+  TRACE_IGNORE_FILES = Dir[File.join(File.dirname(__FILE__), '**', '*.rb')]
+end