+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Performance improvement: Don't trace while in the Pry console. Only works in
   >= 1.9.3-p0 because 1.9.2 segfaults: http://redmine.ruby-lang.org/issues/3921
+* Always cleanup pry-remote DRb server and trace function when a program
+  ends. Fixes [#1](https://github.com/nixme/pry-nav/issues/1).
 
 ## 0.0.2 (2011-11-30)